agent/requirements.txt
livekit-agents[openai]>=0.8
livekit-plugins-openai>=0.8
livekit-plugins-silero>=0.8
